Move _SSB192Selection into pipeline_utils and setup v2
